role:tertaker: Add legacy scripts from old gem
These scripts should be refactored into Jenkins' Job DSL and a Groovy
script to avoid the dependency on this Rubygem (and Ruby).

We can still use ruby for scripting but nowadays we can use more
elaborated techniqes as the Job DSL to create and manage the jobs.

Groovy can take care of the XML dance (which is currently done by
`nokogiri`).

Goal: Simplify the toolset and use tools that are already there (like
Groovy and Job DSL) instead of re-inventing the wheel.