jenkins: Add configuration
This helps us to restore or configure the service.

The goal is to store all configuration inside `files/config`
and this folder get's synced with JENKINS_HOME.

This means we have all configuration on one place.

*DETAILS*

- Rename `jenkins` to `jenkins-remote`

- Move scriptler into `templates/config`

This makes it easier for us to manage the configuration of all plugins in one place.

- Always `backup` config files

The idea is to create a backup before writing the new files.

This can prevent damage to the service in cases where you have invalid
or outdated files on your local machine.

- Store pmd_ruleset inside `config`

This file sets the PMD rules and is used by https://ci.typo3.org/job/typo3-v4-core-nightly/

*CLEANUP*

We also cleanup configuration files to make them easier to understand:

- Use templates and variables for configuration
- Add `defaults` file for service
- Add plugin configuration files
- jenkins: Update JAVA_ARGS to fix upload size
  because encountered an exception when creating new views.

  The new JAVA_ARGS should prevent that.
  See http://stackoverflow.com/q/3861455 for context

- Replace default `all` view with a dashboard
  This loads much quicker than the original view
  and provides a better overview

- Remove slave
  We use the `priority` feature for this build

- Use local copy of logo

- Cleanup default variables for Plugins and Version

- jenkins: Add handlers for safe restart / reload

The idea is to use the CLI for restarting / reloading
which is more sensitive than the init scripts.

The reason for that is that the CLI knows about the
internal state of Jenkins whereas the init script does
not know anything about running jobs etc.

By using the CLI we can avoid hard shutdowns
in the middle of a build for example.