common: Convert roles to `common` task files
We convert the 'heavy' roles to more lightweigt 'tasks'
to reduce the amount of code and to improve maintainability.

Roles introduce a certain amount of overhead that is not needed at the
moment.

Portability of the roles is not a criteria for now. It is more important
to make the code less scary for people that are new to Ansible.

From my experience tasks are easier to understand than roles because we
don't have to deal with seperate README and meta files.

This can be restructured if we feel the need later.

*Details*

- Merge role `ssl`

The SSL config is the same for all servers so `common` seems the place
to be for now.

- Merge `java` tasks

We need java on all servers so putting it into `common` seems to make sense.

- Downgrade to java-1.6 as Version 1.7 is not supported by OS.

- Add packages needed for ansible

- Use ansible modules for locale handling

- Cleanup tasks

 + Add `ssl` and `git` settings
 + Ignore change messages in locales setup
 + Install `python-mysqldb` for mysql module