Vagrantfile: Improve local testing environment
- Support boxes for multiple OS

The reason for this is that we currently need to support two OS because
we migrate the services from the dkd server to an official TYPO3 server:

- lucid64.box = dkd server (Ubuntu)
- squeeze.box = TYPO3 server (Debian)

To fullfill this requirement you can download the lucid64.box from
the vagrant website and store it in vendor.

When the migration is done we will only support the debian box.

- make ansible output less verbose

- Use static IP address for server

This allows us to test the vhost config(s)